feat: create navigation and refactor code
